move statModel to common and some clean up
